Separate FindRunnerService from Runner
